Move supabase and zod to shorter path
