--- src/types/zod/index.ts
@@ -1,1 +1,0 @@
-export * from "@/types/zod/login-form";
