+++ src/zod/user.ts
@@ -1,0 +1,11 @@
+import { z } from "zod";
+
+export const basicUserInfo = z.object({
+  firstName: z.string().nullable(),
+  lastName: z.string().nullable(),
+  email: z.string(),
+  avatarUrl: z.string().url().nullable(),
+  isProfileComplete: z.boolean(),
+});
+
+export type BasicUserInfo = z.infer<typeof basicUserInfo>;
